requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
langchain-groq